+++ client/src/components/Container/basket/BasketContainer.vue
@@ -20,7 +20,7 @@
 </template>
 <script>
 // Stores
-import { getters } from '@/stores/baskets'
+import { useBasketStore } from '@/stores/baskets'
 import { useUserStore } from '@/stores/user'
 // Components
 import Container from '../Container.vue'
@@ -40,16 +40,18 @@ export default {
     title: { type: String, default: 'dashboard.pickupdates' },
   },
   setup () {
+    const basketStore = useBasketStore()
     return {
       userStore,
+      basketStore,
     }
   },
   computed: {
     radius () {
-      return getters.getRadius()
+      return this.basketStore.getRadius
     },
     data () {
-      const data = getters.getNearby()
+      const data = this.basketStore.getNearby()
       this.setList(data)
       return data
     },