+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Refactoring
 - Get rid of xhr completely !3826 @chriswalg
 - Rework working group rest controller !3879 @chriswalg
+- Moved basket store to pinia !3687 @chriswalg
 - Remove jquery-ui and some other jquery extentions !3883 @chriswalg
 - Rework foodsaver rest controller !3892 @chriswalg